+++ services/worker/src/graphview_worker/__init__.py
@@ -1,0 +1,5 @@
+"""Graphview worker service."""
+
+from graphview_worker.pipeline import WorkerStage, build_stage_plan
+
+__all__ = ["WorkerStage", "build_stage_plan"]